feat: introduce `Address` class
